requirements/production.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements/production.in
#
amqp==5.0.9
    # via kombu
arrow==1.2.2
    # via -r requirements/production.in
asgiref==3.5.0
    # via django
attrs==21.4.0
    # via
    #   jsonschema
    #   pytest
billiard==3.6.4.0
    # via celery
bleach==4.1.0
    # via -r requirements/production.in
boto3==1.21.3
    # via django-storages
botocore==1.24.3
    # via
    #   boto3
    #   s3transfer
celery[redis]==5.2.3
    # via
    #   -r requirements/production.in
    #   django-celery-beat
cffi==1.15.0
    # via cryptography
click==8.0.4
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
colorama==0.4.4
    # via rich
commonmark==0.9.1
    # via rich
crispy-bootstrap5==0.6
    # via -r requirements/production.in
cryptography==36.0.1
    # via django-rest-knox
deprecated==1.2.13
    # via redis
django==4.0.2
    # via
    #   -r requirements/production.in
    #   crispy-bootstrap5
    #   django-appconf
    #   django-celery-beat
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   django-health-check
    #   django-redis
    #   django-rest-knox
    #   django-storages
    #   django-timezone-field
    #   djangorestframework
    #   drf-spectacular
django-appconf==1.0.5
    # via django-imagekit
django-celery-beat @ git+https://github.com/celery/django-celery-beat.git@716a46ed34dc64c179fcf022b2aa6e7db146c947
    # via -r requirements/production.in
django-cors-headers==3.11.0
    # via -r requirements/production.in
django-crispy-forms==1.14.0
    # via
    #   -r requirements/production.in
    #   crispy-bootstrap5
django-extensions==3.1.5
    # via -r requirements/production.in
django-filter==21.1
    # via -r requirements/production.in
django-health-check==3.16.5
    # via -r requirements/production.in
django-imagekit==4.1.0
    # via -r requirements/production.in
django-redis==5.2.0
    # via -r requirements/production.in
django-rest-knox==4.2.0
    # via -r requirements/production.in
django-storages[boto3]==1.12.3
    # via -r requirements/production.in
django-timezone-field==4.2.3
    # via django-celery-beat
djangorestframework==3.13.1
    # via
    #   -r requirements/production.in
    #   django-rest-knox
    #   drf-spectacular
drf-spectacular==0.21.2
    # via -r requirements/production.in
execnet==1.9.0
    # via pytest-xdist
factory-boy==3.2.1
    # via -r requirements/production.in
faker==13.0.0
    # via factory-boy
inflection==0.5.1
    # via drf-spectacular
iniconfig==1.1.1
    # via pytest
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
jsonschema==4.4.0
    # via drf-spectacular
kombu==5.2.3
    # via celery
mistune==0.8.4
    # via -r requirements/production.in
packaging==21.3
    # via
    #   bleach
    #   pytest
    #   pytest-sugar
    #   redis
pilkit==2.0
    # via django-imagekit
pillow==9.0.1
    # via -r requirements/production.in
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.28
    # via click-repl
psycopg2-binary==2.9.3
    # via -r requirements/production.in
py==1.11.0
    # via
    #   pytest
    #   pytest-forked
pycparser==2.21
    # via cffi
pygments==2.11.2
    # via rich
pyparsing==3.0.7
    # via packaging
pyrsistent==0.18.1
    # via jsonschema
pytest==7.0.1
    # via
    #   -r requirements/production.in
    #   pytest-django
    #   pytest-forked
    #   pytest-sugar
    #   pytest-xdist
pytest-django==4.5.2
    # via -r requirements/production.in
pytest-forked==1.4.0
    # via pytest-xdist
pytest-sugar==0.9.4
    # via -r requirements/production.in
pytest-xdist==2.5.0
    # via -r requirements/production.in
python-crontab==2.6.0
    # via django-celery-beat
python-dateutil==2.8.2
    # via
    #   arrow
    #   botocore
    #   faker
    #   python-crontab
pytz==2021.3
    # via
    #   celery
    #   django-timezone-field
    #   djangorestframework
pyuwsgi==2.0.20
    # via -r requirements/production.in
pyyaml==6.0
    # via drf-spectacular
redis==4.1.4
    # via
    #   celery
    #   django-redis
rich==11.2.0
    # via -r requirements/production.in
s3transfer==0.5.1
    # via boto3
six==1.16.0
    # via
    #   bleach
    #   click-repl
    #   django-imagekit
    #   python-dateutil
sqlparse==0.4.2
    # via django
termcolor==1.1.0
    # via pytest-sugar
tomli==2.0.1
    # via pytest
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.8
    # via botocore
uwsgitop==0.11
    # via -r requirements/production.in
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via bleach
wrapt==1.13.3
    # via deprecated

# The following packages are considered to be unsafe in a requirements file:
# setuptools
